move go cjs files and export them
