+++ wasm/wasm_exec_node.cjs
